feat: added /datasets/[id] delete endpoint
