+++ sites/geohub/src/lib/server/helpers/index.ts
@@ -16,3 +16,4 @@ export * from './generateHashKey'
 export * from './generateAzureBlobSasToken'
 export * from './upsertDataset'
 export * from './getMe'
+export * from './getDatasetById'
